Refactoring - rename methods and parameters. Update comments Change test
