+++ hardhat.config.js
@@ -270,7 +270,7 @@ task('fee-policy-upgrade-router', 'Updates router to support fee policy logic')
     .addParam('feeCalculatorFacet', 'The address of the FeeCalculatorFacet contract')
     .setAction(async (taskArgs) => {
         console.log(taskArgs);
-        const { upgradeRouter } = require('./scripts/fee-policy-management');
+        const { upgradeRouter } = require('./scripts/fee-policy-management-upgrade');
 
         await upgradeRouter(taskArgs.router, taskArgs.routerFacet, taskArgs.feeCalculatorFacet);
     });
@@ -279,7 +279,7 @@ task('fee-policy-deploy-flat-fee', 'Deploys flat fee policy contract')
     .addParam('flatFee', 'Flat fee value')
     .setAction(async (taskArgs) => {
         console.log(taskArgs);
-        const { deployFlatFeePolicy } = require('./scripts/fee-policy-management');
+        const { deployFlatFeePolicy } = require('./scripts/fee-policy-management-upgrade');
 
         await deployFlatFeePolicy(taskArgs.flatFee);
     });
@@ -289,7 +289,7 @@ task('fee-policy-update-flat-fee', 'Updates flat fee policy contract')
     .addParam('flatFee', 'Flat fee value')
     .setAction(async (taskArgs) => {
         console.log(taskArgs);
-        const { setFlatFeePolicy } = require('./scripts/fee-policy-management');
+        const { setFlatFeePolicy } = require('./scripts/fee-policy-management-upgrade');
 
         await setFlatFeePolicy(taskArgs.feePolicy, taskArgs.flatFee);
     });
@@ -299,7 +299,7 @@ task('fee-policy-deploy-percentage-fee', 'Deploys percentage fee policy contract
     .addParam('feePercentage', 'Fee percentage value')
     .setAction(async (taskArgs) => {
         console.log(taskArgs);
-        const { deployPercentageFeePolicy } = require('./scripts/fee-policy-management');
+        const { deployPercentageFeePolicy } = require('./scripts/fee-policy-management-upgrade');
 
         await deployPercentageFeePolicy(taskArgs.precision, taskArgs.feePercentage);
     });
@@ -310,7 +310,7 @@ task('fee-policy-update-percentage-fee', 'Updates percentage fee policy contract
     .addParam('feePercentage', 'Fee percentage value')
     .setAction(async (taskArgs) => {
         console.log(taskArgs);
-        const { setPercentageFeePolicy } = require('./scripts/fee-policy-management');
+        const { setPercentageFeePolicy } = require('./scripts/fee-policy-management-upgrade');
 
         await setPercentageFeePolicy(taskArgs.feePolicy, taskArgs.precision, taskArgs.feePercentage);
     });
@@ -318,7 +318,7 @@ task('fee-policy-update-percentage-fee', 'Updates percentage fee policy contract
 task('fee-policy-deploy-flat-fee-per-token', 'Deploys flat fee per token policy')
     .setAction(async (taskArgs) => {
         console.log(taskArgs);
-        const { deployFlatFeePerTokenPolicy } = require('./scripts/fee-policy-management');
+        const { deployFlatFeePerTokenPolicy } = require('./scripts/fee-policy-management-upgrade');
 
         await deployFlatFeePerTokenPolicy();
     });
@@ -328,7 +328,7 @@ task('fee-policy-update-flat-fee-per-token', 'Updates flat fee of token')
     .addParam('flatFee', 'Flat fee value')
     .setAction(async (taskArgs) => {
         console.log(taskArgs);
-        const { setFlatFeePerTokenPolicy } = require('./scripts/fee-policy-management');
+        const { setFlatFeePerTokenPolicy } = require('./scripts/fee-policy-management-upgrade');
 
         await setFlatFeePerTokenPolicy(taskArgs.token, taskArgs.flatFee);
     });
@@ -337,7 +337,7 @@ task('fee-policy-remove-flat-fee-per-token', 'Removes token from policy')
     .addParam('token', 'Address of the token')
     .setAction(async (taskArgs) => {
         console.log(taskArgs);
-        const { removeFlatFeePerTokenPolicy } = require('./scripts/fee-policy-management');
+        const { removeFlatFeePerTokenPolicy } = require('./scripts/fee-policy-management-upgrade');
 
         await removeFlatFeePerTokenPolicy(taskArgs.token);
     });
@@ -348,7 +348,7 @@ task('fee-policy-set-users-to-policy', 'Sets fee policy address to user')
     .addParam('addresses', 'User addresses which will be subject of the fee policies separated by comma (",")')
     .setAction(async (taskArgs) => {
         console.log(taskArgs);
-        const { setUsersFeePolicy } = require('./scripts/fee-policy-management');
+        const { setUsersFeePolicy } = require('./scripts/fee-policy-management-upgrade');
 
         await setUsersFeePolicy(taskArgs.router, taskArgs.feePolicy, taskArgs.addresses.split(','));
     });
@@ -358,7 +358,7 @@ task('fee-policy-remove-users-from-policy', 'Removes users from policy')
     .addParam('addresses', 'User addresses which will be subject of the fee policies separated by comma (",")')
     .setAction(async (taskArgs) => {
         console.log(taskArgs);
-        const { removeUsersFeePolicy } = require('./scripts/fee-policy-management');
+        const { removeUsersFeePolicy } = require('./scripts/fee-policy-management-upgrade');
 
         await removeUsersFeePolicy(taskArgs.router, taskArgs.addresses.split(','));
     });
